improved documentation, refactored socket wrapper into its own file
